rework integration tests
